split arbitrary and gen.
